hackingtool/requirements.txt
requests
rich
prompt_toolkit
python-nmap
colorama
python-dotenv
